+++ packages/server/src/services/agents/steward/ops.ts
@@ -3,7 +3,7 @@ import { Observable, map, mergeMap } from 'rxjs'
 import { HexString, NetworkURN } from '@/lib.js'
 import { IngressConsumer } from '@/services/ingress/index.js'
 
-import { asSerializable } from '../base/util.js'
+import { asSerializable } from '@/common/util.js'
 import { AssetMetadata, StorageCodecs, WithRequired } from './types.js'
 import { getLocationIfAny } from './util.js'
