+++ packages/server/src/services/agents/xcm/agent.ts
@@ -4,8 +4,8 @@ import { Operation } from 'rfc6902'
 import { Observable, filter, from, map, switchMap } from 'rxjs'
 import { z } from 'zod'
 
+import { ControlQuery } from '@/common/index.js'
 import { ValidationError } from '@/errors.js'
-import { ControlQuery } from '@/rx/index.js'
 import { getChainId, getConsensus } from '@/services/config.js'
 import { Egress } from '@/services/egress/hub.js'
 import { IngressConsumer } from '@/services/ingress/index.js'
