+++ packages/server/src/services/agents/xcm/ops/ump.ts
@@ -1,13 +1,12 @@
 import { Binary } from 'polkadot-api'
 import { Observable, bufferCount, filter, map, mergeMap } from 'rxjs'
 
+import { asSerializable, filterNonNull } from '@/common/index.js'
 import { HexString } from '@/lib.js'
-import { filterNonNull } from '@/rx/index.js'
 import { getChainId, getRelayId } from '@/services/config.js'
 import { ApiContext, BlockEvent } from '@/services/networking/index.js'
 import { NetworkURN } from '@/services/types.js'
 
-import { asSerializable } from '../../base/util.js'
 import { GetOutboundUmpMessages } from '../types-augmented.js'
 import {
   GenericXcmInboundWithContext,
