+++ packages/server/src/services/subscriptions/switchboard.ts
@@ -3,11 +3,11 @@ import EventEmitter from 'node:events'
 import { Operation } from 'rfc6902'
 
 import { NotFound } from '@/errors.js'
-import { Logger, Services } from '@/services/types.js'
 import { AgentCatalog, AgentId } from '@/services/agents/types.js'
 import { PublisherEvents } from '@/services/egress/types.js'
 import { SubsStore } from '@/services/persistence/level/index.js'
 import { TelemetryCollect, TelemetryEventEmitter } from '@/services/telemetry/types.js'
+import { Logger, Services } from '@/services/types.js'
 
 import { EgressListener, NewSubscription, Subscription, SubscriptionStats } from './types.js'
